alert/requirements.txt
# alert/requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.40
PyMySQL==1.1.0
python-dotenv==1.1.0
pydantic==2.11.4
uvicorn==0.34.2